requirements.txt
fastapi
uvicorn
pandas
duckdb
python-multipart
requests
dotenv
google-genai
